requirements.txt
isodate
yt-dlp 
langdetect
beautifulsoup4 
transformers
pinecone
langchain
langchain-community
dotenv
torch 
psycopg2-binary 
sentence-transformers
sentencepiece
accelerate
langchain-pinecone 
langchain-huggingface
git+https://github.com/openai/whisper.git
language-tool-python
ffmpeg
stanza